build: Moved to sh instead of bash, and use shebang
As build.sh is pretty simple, using a strictly POSIX-compliant shell
(`sh`) is enough. As `bash` isn't available on all systems, it's better
to use `sh`, which is available on all unixes.

This commit adds a shebang to specify the shell to use for build.sh, and
markes it as executable so that you can build by just typing ./build.sh.
This commit also updates the documentation to reflect that change.